+++ src/app/FakeLib/EnvironmentHelper.fs
@@ -17,12 +17,12 @@ type EnvironTarget = EnvironmentVariableTarget
 /// Retrieves the environment variable with the given name
 let environVar name = Environment.GetEnvironmentVariable name
 
-/// Combines two path strings using Path.Combine
+/// Combines two path strings using Path.Combine after removing leading slashes from the second path
 let inline combinePaths path1 (path2 : string) = Path.Combine(path1, path2.TrimStart [| '\\'; '/' |])
 /// Combines two path strings using Path.Combine
 let inline combinePathsNoTrim path1 path2 = Path.Combine(path1, path2)
 
-/// Combines two path strings using Path.Combine
+/// Combines two path strings using Path.Combine after removing leading slashes from the second path
 let inline (@@) path1 path2 = combinePaths path1 path2
 /// Combines two path strings using Path.Combine
 let inline (</>) path1 path2 = combinePathsNoTrim path1 path2
@@ -173,8 +173,11 @@ let mutable TargetPlatformPrefix =
         else Some @"C:\Windows\Microsoft.NET\Framework"
     |> Option.get
 
-/// Base path for getting tools from windows SDKs
-let sdkBasePath = ProgramFilesX86 @@ "Microsoft SDKs\Windows"
+/// Base path for getting tools from Microsoft SDKs
+let msSdkBasePath = ProgramFilesX86 @@ "Microsoft SDKs"
+
+/// Base path for getting tools from Windows SDKs
+let sdkBasePath = msSdkBasePath @@ "Windows"
 
 /// Helper function to help find framework or sdk tools from the 
 /// newest toolkit available