+++ src/app/FakeLib/NpmHelper.fs
@@ -36,12 +36,19 @@ type InstallArgs =
 | Standard
 | Forced
 
-/// The list of supported Npm commands. The `Custom` alternative
-/// can be used for other commands not in the list until they are
-/// implemented
+/// The list of supported Npm commands.
 type NpmCommand =
+/// Run `npm install`
 | Install of InstallArgs
+/// Run `npm run <string>`
 | Run of string
+/// Run `npm run --silent <string>`. Suppresses npm error output. See [npm:8821](https://github.com/npm/npm/issues/8821).
+| RunSilent of string
+/// Run `npm run --silent <string>`. Suppresses npm error output and will raise an FailedTestsException exception after the script execution instead of failing, useful for CI. See [npm:8821](https://github.com/npm/npm/issues/8821).
+| RunTest of string
+/// Run `npm test --silent`. Suppresses npm error output and will raise an FailedTestsException exception after the script execution instead of failing, useful for CI. See [npm:8821](https://github.com/npm/npm/issues/8821).
+| Test
+/// Run `npm <string>`. Can be used for running not implemented commands.
 | Custom of string
 
 /// The Npm parameter type
@@ -66,19 +73,36 @@ let private parseInstallArgs = function
     | Forced -> " --force"
 
 let private parse = function
-    | Install installArgs -> sprintf "install%s" (installArgs |> parseInstallArgs)
+    | Install installArgs -> sprintf "install %s" (installArgs |> parseInstallArgs)
     | Run str -> sprintf "run %s" str
+    | RunSilent str -> sprintf "run --silent %s" str
+    | RunTest str -> sprintf "run --silent %s" str
     | Custom str -> str
+    | Test -> "test --silent"
 
+/// Runs the given process and returns the process result.
 let run npmParams =
+    let result = ref None
     let npmPath = Path.GetFullPath(npmParams.NpmFilePath)
-    let arguments = npmParams.Command |> parse
-    let ok = 
-        execProcess (fun info ->
-            info.FileName <- npmPath
-            info.WorkingDirectory <- npmParams.WorkingDirectory
-            info.Arguments <- arguments) npmParams.Timeout
-    if not ok then failwith (sprintf "'npm %s' task failed" arguments)
+    let args = npmParams.Command |> parse
+    try 
+        let exitCode = 
+            ExecProcess (fun info -> 
+                info.WorkingDirectory <- npmParams.WorkingDirectory
+                info.FileName <- npmPath
+                info.Arguments <- args) npmParams.Timeout
+        if exitCode <> 0 then result := Some(sprintf "exit code: %d" exitCode)
+    with exn ->
+        let message = ref exn.Message
+        if exn.InnerException <> null then message := !message + Environment.NewLine + exn.InnerException.Message
+        result := Some(!message)
+    match !result with
+    | None -> ()
+    | Some msg ->
+        match npmParams.Command with
+        | RunTest str -> raise (UnitTestCommon.FailedTestsException("Test(s) Failed"))
+        | Test -> raise (UnitTestCommon.FailedTestsException("Test(s) Failed"))
+        | _ -> failwith msg
 
 /// Runs npm with the given modification function. Make sure to have npm installed,
 /// you can install npm with nuget or a regular install. To change which `Npm` executable
@@ -103,5 +127,4 @@ let run npmParams =
 ///                       WorkingDirectory = "./src/FAKESimple.Web/"
 ///                   })
 ///        )
-let Npm setParams =
-    defaultNpmParams |> setParams |> run
+let Npm setParams = defaultNpmParams |> setParams |> run